library/src/main/java/ua/nykyforov/service/library/application/service/AuthorServiceImpl.java
package ua.nykyforov.service.library.application.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import ua.nykyforov.service.library.core.application.AuthorService;
import ua.nykyforov.service.library.core.dao.AuthorDao;
import ua.nykyforov.service.library.core.domain.Author;

import javax.transaction.Transactional;
import java.util.Optional;

@Service
public class AuthorServiceImpl implements AuthorService {

    private final AuthorDao authorDao;

    @Autowired
    public AuthorServiceImpl(AuthorDao authorDao) {
        this.authorDao = authorDao;
    }

    @Override
    @Transactional
    public void save(Author author) {
        authorDao.insert(author);
    }

    @Override
    @Transactional
    public Optional<Author> getById(int id) {
        return Optional.ofNullable(authorDao.getById(id));
    }

}
